version check function : --version_input=???

+++ 1_excel.pl
@@ -20,13 +20,18 @@ use MY::CHARLES qw(traverse_tree_to_file);
 use MY::CHARLES qw(traverse_hash_tree);
 
 use Getopt::Long;
-my $input_file   = "./1_example.xlsx";
-my $output_file = "./default.GV";
-my $csv_file = "./1_example.csv";
-my $help=0;
-my $verbose;
+our $input_file   = "./1_example.xlsx";
+our $version_input_file   = "";
+our $output_file = "./default.GV";
+our $csv_file = "./1_example.csv";
+our $help=0;
+our $verbose;
+our $verMajor;
+our $verMinor;
+our $verDev;
 GetOptions ("output=s" => \$output_file,    # string
 		"input=s"   => \$input_file,      # string
+		"version_input=s"   => \$version_input_file,      # string
 		"csv_output=s"   => \$csv_file,      # string
 		"verbose|help"  => sub { $help = 1 })   # flag
 or  die(help() . "Error in command line arguments\n");
@@ -34,6 +39,19 @@ or  die(help() . "Error in command line arguments\n");
 if($help == 1){
 	help();
 	exit();
+}
+
+print STDERR "version file : $version_input_file\n";
+
+if($version_input_file ne ""){
+    open(my $verfh, "<", $version_input_file);
+    my $s = <$verfh>;
+    $s =~ /\s*v\s*(\d+)\.(\d+)\.(\d+)\s*/;
+    $verMajor = $1;
+    $verMinor = $2;
+    $verDev = $3;
+    print STDERR "CURRENT VERSION $s => Major $verMajor, Minor $verMinor, Dev $verDev\n";
+    close($verfh);
 }
 
 print "[" . abs_path . "]\n";
@@ -184,6 +202,21 @@ say '[1] maxrow:' . $book->[1]{maxrow};
 
 hashTraverseSTDOUT();
 
+if($version_input_file ne ""){
+    my $tmpMajor = 0;
+    my $tmpMinor = 0;
+    my $tmpDev = 0;
+    my $s = $gTitle{"VARIABLE"}{"Excel_Version"};
+    $s =~ /\s*v\s*(\d+)\.(\d+)\.(\d+)\s*/;
+    $tmpMajor = $1;
+    $tmpMinor = $2;
+    $tmpDev = $3;
+    print STDERR "EXCEL FILE VERSION $s => Major $tmpMajor, Minor $tmpMinor, Dev $tmpDev\n";
+    if($verMajor != $tmpMajor) { versionMismatch(); exit(4); }
+    if($verMinor != $tmpMinor) { versionMismatch(); exit(4); }
+    if($verDev != $tmpDev) { versionMismatch(); exit(4); }
+}
+
 # $gPrintHashName{"LXID_ROTATE"} = "telltale of type ROTATE";
 unlink $output_file;  # or warn "Could not unlink $output_file: $!";
 traverse_tree_to_file(\%gTitle,"gTitle",">>",$output_file);
@@ -197,6 +230,13 @@ foreach $mn (keys %Module_Name){
      or die "Can't open > $lmn : $!";
     print $mnfh "$csv_file";
     close($mnfh);
+}
+
+sub versionMismatch {
+	print STDERR "Version Mismatch between excel file and this git repository\n";
+	print STDERR "1.check Excel_Version in excel file\n";
+	print STDERR "2.check Version in $version_input_file file\n";
+	print STDERR "Please Convert your excel file into latest excel version.\n";
 }
 
 sub FixXML {
@@ -275,6 +315,9 @@ sub help
 	printf("Help :\n");
 	printf("\t--input=[input excel file]\n");
 	printf("\t\t  default input file name : $input_file\n");
+	printf("\t--version_input=[version input file]\n");
+	printf("\t\t  default version input file name : $version_input_file\n");
+	printf("\t\t  if null , we do not check the version between version input file and [VARIABLE]Excel_Version  in excel file ($input_file)\n");
 	printf("\t--output=[output file with global variables]\n");
 	printf("\t\t  default output file name : $output_file\n");
 	printf("\t--csv_out=[csv output file]\n");
